requirements.txt
arrow==1.2.3
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
icalendar==5.0.4
ics==0.7.2
idna==3.4
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
TatSu==5.8.3
urllib3==1.26.14
